requirements.txt
langchain_core==1.2.5
langchain_openai==1.1.6
pandas==2.3.3
python-dotenv==1.2.1
streamlit==1.52.0
openpyxl==3.1.5